+++ docs/changelog.md
@@ -15,7 +15,10 @@ pageflow_next_text: Contributing
 * Added support for Checks
   - added `Check` interface
   - added `ListCheck` interface
+  - added `BadCheck` exception
   - added `BadCheckArgs` exception
+  - added `BadChecksList` exception
+  - added `EmptyChecksList` exception
   - added `ComposableCheck`
 
 ### Fixes